HRHiringSystem.AIAgent/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12

# Google Gemini AI
google-generativeai==0.8.2

# LangChain (without OpenAI)
langchain==0.3.4
langchain-community==0.3.3

# Document parsing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.11.0

# Azure
azure-storage-blob==12.23.0

# HTTP client
httpx==0.27.2
aiohttp==3.10.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
